Move dos/bands routine from separate to one fleur module in vis (Makes it easier to find)
Also remove some "decoy" files, that are currently empty, we can just recreate them if we need it
