--- masci_tools/vis/base_plots.py
@@ -1,15 +1,0 @@
-# -*- coding: utf-8 -*-
-###############################################################################
-# Copyright (c), Forschungszentrum Jülich GmbH, IAS-1/PGI-1, Germany.         #
-#                All rights reserved.                                         #
-# This file is part of the Masci-tools package.                               #
-# (Material science tools)                                                    #
-#                                                                             #
-# The code is hosted on GitHub at https://github.com/judftteam/masci-tools.   #
-# For further information on the license, see the LICENSE.txt file.           #
-# For further information please visit http://judft.de/.                      #
-#                                                                             #
-###############################################################################
-"""
-Basic plots that wrap matplotlib plots
-"""
